Ai4Green_Api/pyproject.toml
[tool.poetry]
name = "ai4green"
version = "1.5.1"
description = "AI4Green is an Electronic Laboratory Notebook which combines data storage and sharing while promoting green and sustainable chemistry."
authors = ["<ai4green@nottingham.ac.uk>"]
license = "AGPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
azure-core = "^1.29.3"
azure-identity = "^1.14.0"
azure-storage-blob = "==12.17.0"
beautifulsoup4 = "^4.12.2"
chemspipy = "^2.0.0"
click = "^8.1.7"
email-validator = "^2.0.0.post2"
flask = "^2.3.3"
flask-login = "^0.6.2"
flask-mail = "^0.9.1"
flask-marshmallow = "^0.15.0"
flask-migrate = "^4.0.4"
flask-sqlalchemy = "^3.0.5"
flask-wtf = "^1.1.1"
lxml = "^4.9.3"

marshmallow-sqlalchemy = "^0.29.0"
pandas = "^2.0.3"
pikachu-chem = "^1.0.10"
plotly = "^5.16.1"
psycopg2-binary = "^2.9.7"
puremagic = "^1.15"
pyyaml = "^6.0.1"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"
python-magic = "^0.4.27"
pytz = "^2023.3"
rdkit = "^2023.3.3"
requests = "^2.31.0"
wtforms = "^3.0.1"
toml = "^0.10.2"
numpy = "^1.24.3"
werkzeug = "==2.3.7"

dash = "^2.6.2"
dash-cytoscape = "^0.3.0"
dash-bootstrap-components = "^1.2.1"
dash-html-components = "^2.0.0"
dash-table = "^5.0.0"
dash-core-components = "^2.0.0"
#pikachu-chem = "^1.1.0"
odfpy = "^1.4.1"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
selenium = "^4.11.2"
flask-testing = "^0.8.1"
coverage = "^7.3.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
isort = "^5.12.0"
mypy = "^1.5.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "Webapp/tests",
]
